Add files via upload
Spatial data require to be graphed in a scheme of colors that make it more clear and understandable. So, if you would like to do more informative plotting your spatial data using Matplotlip #Python library, logarithmic scale does a great job. Here is an example that shows you the different between normal and #logarithmic_scale plotting using an attribute of  data.